requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
python-dotenv==1.0.1
aiohttp==3.11.11
tronpy==0.5.0
